@@ -9,8 +9,9 @@
     "type": "git",
     "url": "https://github.com/GoogleCloudPlatform/nodejs-docs-samples.git"
   },
+  "type": "module",
   "engines": {
-    "node": ">=16.0.0"
+    "node": ">=20.0.0"
   },
   "main": "main.js",
   "scripts": {
@@ -23,12 +24,12 @@
   "dependencies": {
     "express": "^4.17.1",
     "google-auth-library": "^9.0.0",
-    "got": "^11.5.0",
-    "handlebars": "^4.7.6"
+    "got": "^14.6.5",
+    "handlebars": "^4.7.8"
   },
   "devDependencies": {
     "c8": "^10.0.0",
-    "mocha": "^10.0.0",
+    "mocha": "^11.0.0",
     "supertest": "^7.0.0"
   }
 }
